# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FC Recruits - EA Sports FC Pro Clubs recruitment platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  userId: string;
  gamertag: string;
  /** @nullable */
  displayName?: string | null;
  /** PS5, Xbox, PC */
  platform: string;
  /** GK, RB, CB, LB, CDM, CM, CAM, RM, LM, RW, LW, ST */
  mainPosition: string;
  /** @nullable */
  secondaryPosition?: string | null;
  /** Player OVR 1-99 */
  overallRating: number;
  /** @nullable */
  goals?: number | null;
  /** @nullable */
  assists?: number | null;
  /** @nullable */
  passAccuracy?: number | null;
  /** @nullable */
  cleanSheets?: number | null;
  /** @nullable */
  matchesPlayed?: number | null;
  country: string;
  timezone: string;
  /**
     * JSON array of days e.g. ["Mon","Tue"]
     * @nullable
     */
  availableDays?: string | null;
  /**
     * HH:MM
     * @nullable
     */
  availableFrom?: string | null;
  /**
     * HH:MM
     * @nullable
     */
  availableTo?: string | null;
  /** @nullable */
  bio?: string | null;
  freeAgent: boolean;
  /** @nullable */
  clubId?: number | null;
  /** @nullable */
  clubName?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  averageRating?: number | null;
  /** @nullable */
  totalRatings?: number | null;
  createdAt: string;
}

export interface PlayerList {
  players: Player[];
  total: number;
}

export interface PlayerInput {
  gamertag: string;
  displayName?: string;
  platform: string;
  mainPosition: string;
  secondaryPosition?: string;
  overallRating: number;
  goals?: number;
  assists?: number;
  passAccuracy?: number;
  cleanSheets?: number;
  matchesPlayed?: number;
  country: string;
  timezone: string;
  availableDays?: string;
  availableFrom?: string;
  availableTo?: string;
  bio?: string;
  avatarUrl?: string;
}

export interface PlayerUpdate {
  gamertag?: string;
  displayName?: string;
  platform?: string;
  mainPosition?: string;
  secondaryPosition?: string;
  overallRating?: number;
  goals?: number;
  assists?: number;
  passAccuracy?: number;
  cleanSheets?: number;
  matchesPlayed?: number;
  country?: string;
  timezone?: string;
  availableDays?: string;
  availableFrom?: string;
  availableTo?: string;
  bio?: string;
  freeAgent?: boolean;
  avatarUrl?: string;
}

export interface Club {
  id: number;
  captainId: number;
  /** @nullable */
  captainGamertag?: string | null;
  name: string;
  platform: string;
  /** Division 1-10 (1 = top) */
  division: number;
  /** casual, competitive, semi-competitive */
  style: string;
  country: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /**
     * JSON array of days
     * @nullable
     */
  playDays?: string | null;
  /** @nullable */
  playFrom?: string | null;
  /** @nullable */
  playTo?: string | null;
  /**
     * JSON array of open positions
     * @nullable
     */
  openPositions?: string | null;
  /** @nullable */
  memberCount?: number | null;
  /** @nullable */
  maxMembers?: number | null;
  /** @nullable */
  wins?: number | null;
  /** @nullable */
  losses?: number | null;
  /** @nullable */
  draws?: number | null;
  createdAt: string;
}

export interface ClubList {
  clubs: Club[];
  total: number;
}

export interface ClubInput {
  name: string;
  platform: string;
  division: number;
  style: string;
  country: string;
  description?: string;
  logoUrl?: string;
  playDays?: string;
  playFrom?: string;
  playTo?: string;
  openPositions?: string;
  maxMembers?: number;
}

export interface ClubUpdate {
  name?: string;
  platform?: string;
  division?: number;
  style?: string;
  country?: string;
  description?: string;
  logoUrl?: string;
  playDays?: string;
  playFrom?: string;
  playTo?: string;
  openPositions?: string;
  maxMembers?: number;
  wins?: number;
  losses?: number;
  draws?: number;
}

export interface Member {
  id: number;
  clubId: number;
  playerId: number;
  /** @nullable */
  playerGamertag?: string | null;
  /** @nullable */
  playerPosition?: string | null;
  /** @nullable */
  playerRating?: number | null;
  /** captain, vice-captain, player */
  role: string;
  joinedAt: string;
}

export interface MemberList {
  members: Member[];
}

export interface MemberInput {
  playerId: number;
  role?: string;
}

export interface Tryout {
  id: number;
  clubId: number;
  /** @nullable */
  clubName?: string | null;
  /** @nullable */
  clubLogoUrl?: string | null;
  playerId: number;
  /** @nullable */
  playerGamertag?: string | null;
  /** @nullable */
  playerPosition?: string | null;
  /** @nullable */
  playerRating?: number | null;
  /** player or club */
  initiatedBy: string;
  /** pending, accepted, rejected, scheduled, completed, cancelled */
  status: string;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  message?: string | null;
  /** @nullable */
  playerRatingScore?: number | null;
  /** @nullable */
  clubRatingScore?: number | null;
  createdAt: string;
  updatedAt?: string;
}

export interface TryoutList {
  tryouts: Tryout[];
  total: number;
}

export interface TryoutInput {
  clubId: number;
  playerId: number;
  initiatedBy: string;
  message?: string;
}

export interface TryoutUpdate {
  status?: string;
  scheduledAt?: string;
  message?: string;
}

export interface Rating {
  id: number;
  tryoutId: number;
  /** player or club */
  ratedBy: string;
  /** 1.0 to 5.0 */
  score: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface RatingInput {
  ratedBy: string;
  score: number;
  comment?: string;
}

export interface Notification {
  id: number;
  userId: string;
  /** tryout_request, tryout_accepted, tryout_rejected, tryout_scheduled, tryout_rated, club_invite */
  type: string;
  message: string;
  /** @nullable */
  relatedId?: number | null;
  read: boolean;
  createdAt: string;
}

export interface NotificationList {
  notifications: Notification[];
  unreadCount: number;
}

export interface NotificationUpdate {
  read?: boolean;
}

export interface PlatformBreakdown {
  platform: string;
  count: number;
}

export interface PositionBreakdown {
  position: string;
  count: number;
}

export interface PlatformStats {
  totalPlayers: number;
  totalClubs: number;
  totalTryouts: number;
  freeAgents: number;
  activeClubs: number;
  byPlatform?: PlatformBreakdown[];
  byPosition?: PositionBreakdown[];
}

export type ListPlayersParams = {
position?: string;
platform?: string;
country?: string;
timezone?: string;
minRating?: number;
maxRating?: number;
available?: boolean;
freeAgent?: boolean;
search?: string;
limit?: number;
offset?: number;
};

export type ListClubsParams = {
platform?: string;
style?: string;
division?: number;
openPosition?: string;
country?: string;
search?: string;
limit?: number;
offset?: number;
};

export type ListTryoutsParams = {
role?: string;
status?: string;
playerId?: number;
clubId?: number;
};