# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FC Recruits - EA Sports FC Pro Clubs recruitment platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clubs:
    get:
      operationId: listClubs
      tags: [clubs]
      summary: List and search clubs with filters
      parameters:
        - name: platform
          in: query
          schema:
            type: string
        - name: style
          in: query
          schema:
            type: string
        - name: division
          in: query
          schema:
            type: integer
        - name: openPosition
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of clubs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClubList"
    post:
      operationId: createClub
      tags: [clubs]
      summary: Create a club
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
...
  /clubs/{id}:
    get:
      operationId: getClub
      tags: [clubs]
      summary: Get club by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
        "404":
          description: Not found
    patch:
      operationId: updateClub
      tags: [clubs]
      summary: Update club
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClubUpdate"
      responses:
        "200":
          description: Updated club
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Club"
    delete:
      operationId: deleteClub
      tags: [clubs]
      summary: Delete club
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /clubs/{id}/members:
    get:
      operationId: getClubMembers
      tags: [clubs]
      summary: Get members of a club
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Club members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberList"
    post:
      operationId: addClubMember
      tags: [clubs]
      summary: Add member to club
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /clubs/{id}/members/{playerId}:
    delete:
      operationId: removeClubMember
      tags: [clubs]
      summary: Remove member from club
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: playerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  # ─── TRYOUTS ────────────────────────────────────────────────────────────────
...
  /notifications/{id}:
    patch:
      operationId: updateNotification
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationUpdate"
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ─── STATS ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Club:
      type: object
      required: [id, captainId, name, platform, division, style, country, createdAt]
      properties:
        id:
          type: integer
        captainId:
          type: integer
        captainGamertag:
          type: ["string", "null"]
        name:
          type: string
        platform:
          type: string
        division:
          type: integer
          description: "Division 1-10 (1 = top)"
        style:
          type: string
          description: "casual, competitive, semi-competitive"
        country:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        playDays:
          type: ["string", "null"]
          description: "JSON array of days"
        playFrom:
          type: ["string", "null"]
        playTo:
          type: ["string", "null"]
        openPositions:
          type: ["string", "null"]
          description: "JSON array of open positions"
        memberCount:
          type: ["integer", "null"]
        maxMembers:
          type: ["integer", "null"]
        wins:
          type: ["integer", "null"]
        losses:
          type: ["integer", "null"]
        draws:
          type: ["integer", "null"]
        createdAt:
          type: string
    ClubList:
      type: object
      required: [clubs, total]
      properties:
        clubs:
          type: array
          items:
            $ref: "#/components/schemas/Club"
        total:
          type: integer
    ClubInput:
      type: object
      required: [name, platform, division, style, country]
      properties:
        name:
          type: string
        platform:
          type: string
        division:
          type: integer
        style:
          type: string
        country:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        playDays:
          type: string
        playFrom:
          type: string
        playTo:
          type: string
        openPositions:
          type: string
        maxMembers:
          type: integer
    ClubUpdate:
      type: object
      properties:
        name:
          type: string
        platform:
          type: string
        division:
          type: integer
        style:
          type: string
        country:
          type: string
        description:
          type: string
        logoUrl:
          type: string
        playDays:
          type: string
        playFrom:
          type: string
        playTo:
          type: string
        openPositions:
          type: string
        maxMembers:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        draws:
          type: integer

    # ─── MEMBER SCHEMAS ────────────────────────────────────────────────────────
    Member:
      type: object
      required: [id, clubId, playerId, role, joinedAt]
      properties:
        id:
          type: integer
        clubId:
          type: integer
        playerId:
          type: integer
        playerGamertag:
          type: ["string", "null"]
        playerPosition:
          type: ["string", "null"]
        playerRating:
          type: ["integer", "null"]
        role:
          type: string
          description: "captain, vice-captain, player"
        joinedAt:
          type: string
    MemberList:
      type: object
      required: [members]
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/Member"
    MemberInput:
      type: object
      required: [playerId]
      properties:
        playerId:
          type: integer
        role:
          type: string

    # ─── TRYOUT SCHEMAS ────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, userId, type, message, read, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          description: "tryout_request, tryout_accepted, tryout_rejected, tryout_scheduled, tryout_rated, club_invite"
        message:
          type: string
        relatedId:
          type: ["integer", "null"]
        read:
          type: boolean
        createdAt:
          type: string
...
    NotificationUpdate:
      type: object
      properties:
        read:
          type: boolean

    # ─── STATS SCHEMAS ─────────────────────────────────────────────────────────
...