# lib/api-zod/src/generated/types/club.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FC Recruits - EA Sports FC Pro Clubs recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Club {
  id: number;
  captainId: number;
  /** @nullable */
  captainGamertag?: string | null;
  name: string;
  platform: string;
  /** Division 1-10 (1 = top) */
  division: number;
  /** casual, competitive, semi-competitive */
  style: string;
  country: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /**
     * JSON array of days
     * @nullable
     */
  playDays?: string | null;
  /** @nullable */
  playFrom?: string | null;
  /** @nullable */
  playTo?: string | null;
  /**
     * JSON array of open positions
     * @nullable
     */
  openPositions?: string | null;
  /** @nullable */
  memberCount?: number | null;
  /** @nullable */
  maxMembers?: number | null;
  /** @nullable */
  wins?: number | null;
  /** @nullable */
  losses?: number | null;
  /** @nullable */
  draws?: number | null;
  createdAt: string;
}