# lib/api-zod/src/generated/types/clubList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FC Recruits - EA Sports FC Pro Clubs recruitment platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Club } from './club';

export interface ClubList {
  clubs: Club[];
  total: number;
}