# lib/api-zod/src/generated/types/clubUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FC Recruits - EA Sports FC Pro Clubs recruitment platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ClubUpdate {
  name?: string;
  platform?: string;
  division?: number;
  style?: string;
  country?: string;
  description?: string;
  logoUrl?: string;
  playDays?: string;
  playFrom?: string;
  playTo?: string;
  openPositions?: string;
  maxMembers?: number;
  wins?: number;
  losses?: number;
  draws?: number;
}